Enhance center estimation functions
`center()` now checks the taxon co-occurrence network and throws an
error if there is more than one component---unless the `components =
TRUE` option is set, in which case a data frame that provides the
components assignment along with the center estimate is provided. This
behavior resolves #1 and #2.

`center()` no longer takes the `bound` parameter, since the `rss` method
is not recommended.

The component-checking significantly slowed the `center()` function, and
so `bootrep_center()` now avoids calling `center()` for each replicate,
using the primitive `log_center_*()` functions instead.