Revise and expand mean-efficiency functionality
The interface to `mean_efficiency()` is changed. Instead of the bias
vector, the first argument is now the abundance matrix (or phyloseq
object) representing the actual or observed compositions. Matrices,
phyloseq objects, and mc_bias_fit objects are all supported.

In addition, `calibrate()` now adds the mean efficiency to the sample
data when calibrating a phyloseq object.